refactor: remove deprecated attributes from various XML schemas and related components; update components to improve functionality and documentation
