+++ sdk/longlink/.static/new/src/pages/items.xml
@@ -11,12 +11,8 @@
         <Dialog title="New item" triggerLabel="New item" purpose="form">
             <Stack>
                 <Stack>
-                    <TextInput
-                        label="Name"
-                        value="$draft.name"
-                        isRequired="true"
-                    />
-                    <NumberInput label="Price" units="CHF" value="$draft.price" min="0" isRequired="true" />
+                    <TextInput label="Name" value="$draft.name" />
+                    <NumberInput label="Price" units="CHF" value="$draft.price" min="0" />
                 </Stack>
                 <Action>
                     <Request
