+++ sdk/longlink/.static/xsd/adapters/NumberInput.xsd
@@ -37,11 +37,6 @@
         <xsd:documentation>Increment between allowed values.</xsd:documentation>
       </xsd:annotation>
     </xsd:attribute>
-    <xsd:attribute name="isRequired" type="xsd:string">
-      <xsd:annotation>
-        <xsd:documentation>Marks the field as required.</xsd:documentation>
-      </xsd:annotation>
-    </xsd:attribute>
     <xsd:attribute name="isIntegerOnly" type="xsd:string">
       <xsd:annotation>
         <xsd:documentation>Restricts values to integers.</xsd:documentation>
