+++ sdk/longlink/.static/xsd/adapters/Selector.xsd
@@ -85,26 +85,6 @@
         <xsd:documentation>Shows a search field.</xsd:documentation>
       </xsd:annotation>
     </xsd:attribute>
-    <xsd:attribute name="isRequired" type="xsd:string">
-      <xsd:annotation>
-        <xsd:documentation>Marks the field as required.</xsd:documentation>
-      </xsd:annotation>
-    </xsd:attribute>
-    <xsd:attribute name="isOptional" type="xsd:string">
-      <xsd:annotation>
-        <xsd:documentation>Marks the field as optional.</xsd:documentation>
-      </xsd:annotation>
-    </xsd:attribute>
-    <xsd:attribute name="isLabelHidden" type="xsd:string">
-      <xsd:annotation>
-        <xsd:documentation>Visually hides the label.</xsd:documentation>
-      </xsd:annotation>
-    </xsd:attribute>
-    <xsd:attribute name="isLoading" type="xsd:boolean">
-      <xsd:annotation>
-        <xsd:documentation>Shows a loading state.</xsd:documentation>
-      </xsd:annotation>
-    </xsd:attribute>
     <xsd:attribute name="isDefaultOpen" type="xsd:boolean">
       <xsd:annotation>
         <xsd:documentation>Opens the options layer initially.</xsd:documentation>
