+++ sdk/longlink/.static/xsd/adapters/TextInput.xsd
@@ -22,11 +22,6 @@
         <xsd:documentation>Supporting field text.</xsd:documentation>
       </xsd:annotation>
     </xsd:attribute>
-    <xsd:attribute name="isRequired" type="xsd:string">
-      <xsd:annotation>
-        <xsd:documentation>Marks the field as required.</xsd:documentation>
-      </xsd:annotation>
-    </xsd:attribute>
     <xsd:attributeGroup ref="XmlRuntimeAttributes"/>
   </xsd:complexType>
   <xsd:element name="TextInput" type="TextInputType">
