+++ web/src/xml/adapters/CheckboxInput.tsx
@@ -14,11 +14,7 @@ export function CheckboxInput({ props }: Props) {
     const size = resolveXml(props, 'size', ctx);
     const width = resolveXml(props, 'width', ctx);
     const htmlName = resolveXml(props, 'htmlName', ctx);
-    const isLoading = resolveXml(props, 'isLoading', ctx);
-    const isOptional = resolveXml(props, 'isOptional', ctx);
-    const isRequired = resolveXml(props, 'isRequired', ctx);
     const description = resolveXml(props, 'description', ctx);
-    const isLabelHidden = resolveXml(props, 'isLabelHidden', ctx);
 
     if (!isXmlEnum(size, [undefined, ...COMPACT_SIZES])) {
         throw new Error(`Unsupported CheckboxInput size '${String(size)}'`);
@@ -33,11 +29,7 @@ export function CheckboxInput({ props }: Props) {
             status={resolveInputStatus(props, ctx)}
             htmlName={typeof htmlName === 'string' ? htmlName : undefined}
             onChange={binding.setValue}
-            isLoading={typeof isLoading === 'boolean' ? isLoading : undefined}
-            isOptional={typeof isOptional === 'boolean' ? isOptional : undefined}
-            isRequired={typeof isRequired === 'boolean' ? isRequired : undefined}
             description={typeof description === 'string' ? description : undefined}
-            isLabelHidden={typeof isLabelHidden === 'boolean' ? isLabelHidden : undefined}
         />
     );
 }
