+++ web/src/xml/adapters/TextArea.tsx
@@ -9,7 +9,6 @@ export function TextArea({ props }: Props) {
     const binding = useBindableValue(props, 'value', ctx, (value) => String(value ?? ''));
     const rows = resolveXml(props, 'rows', ctx);
     const maxLength = resolveXml(props, 'maxLength', ctx);
-    const isRequired = resolveXml(props, 'isRequired', ctx);
     const description = resolveXml(props, 'description', ctx);
 
     if (maxLength != null && (typeof maxLength !== 'number' || !Number.isInteger(maxLength) || maxLength < 0)) {
@@ -27,7 +26,6 @@ export function TextArea({ props }: Props) {
             value={binding.value}
             onChange={binding.setValue}
             maxLength={typeof maxLength === 'number' ? maxLength : undefined}
-            isRequired={typeof isRequired === 'boolean' ? isRequired : undefined}
             description={typeof description === 'string' ? description : undefined}
         />
     );
