--- web/src/xml/expressions/index.ts
@@ -1,2 +1,0 @@
-export { compileAttribute } from './compile';
-export { evaluate } from './evaluate';
