+++ web/tests/xml/helpers.tsx
@@ -1,7 +1,7 @@
 import { RenderXML } from '@/xml';
 import { createContext } from '@/xml/core/context';
-import { compileAttribute } from '@/xml/expressions';
 import { renderToStaticMarkup } from 'react-dom/server';
+import { compileAttribute } from '@/xml/expressions/compile';
 import type { ASTNode, ASTProps, XmlRuntime } from '@/xml/types';
 
 /** Compiles string fixture attributes through the same document compiler rules. */
